+++ src/main/java/com/sofodev/sworddisplay/blocks/TESRSwordDisplay.java
@@ -60,17 +60,8 @@ public class TESRSwordDisplay extends TileEntityRenderer<SwordDisplayTile> {
         }
     }
 
-    //         switch (te.getBlockState().get(SwordDisplayBlock.FACING)) {
-    //            case WEST:
-    //            case EAST:
-    //                this.rotateItem(matrixStack, 180f, 90f, 45f); // default vlues
-    //                break;
-    //            case NORTH:
-    //            case SOUTH:
-    //                this.rotateItem(matrixStack, 180f, 180f, 45f);
-    //                break;
-    //        }
-    //
-    //        renderer.renderItem(stack, NONE, combinedLight, combinedOverlay, matrixStack, buffer);
-
+    @Override
+    public boolean isGlobalRenderer(SwordDisplayTile te) {
+        return true;
+    }
 }